docs: remove internal prompt schema machinery
